+++ src/Reader.ts
@@ -1,17 +1,16 @@
 /**
  * @since 2.0.0
  */
+import { Applicative2 } from './Applicative'
+import { Apply2 } from './Apply'
 import { Category2 } from './Category'
 import * as F from './function'
+import { Functor2 } from './Functor'
 import * as I from './Identity'
 import { Monad2 } from './Monad'
 import { Monoid } from './Monoid'
 import { Profunctor2 } from './Profunctor'
-import * as ReaderT from './ReaderT'
 import { Semigroup } from './Semigroup'
-import { Functor2 } from './Functor'
-import { Apply2 } from './Apply'
-import { Applicative2 } from './Applicative'
 import { Semigroupoid2 } from './Semigroupoid'
 
 /**
@@ -49,9 +48,7 @@ export const ask: <R>() => Reader<R, R> = () => I.of
  *
  * @since 2.0.0
  */
-export const asks: <R, A>(f: (r: R) => A) => Reader<R, A> =
-  /*#__PURE__*/
-  ReaderT.asks(I.monadIdentity)
+export const asks: <R, A>(f: (r: R) => A) => Reader<R, A> = (f) => (r) => F.pipe(I.of(r), I.map(f))
 
 /**
  * Changes the value of the local context during the execution of the action `ma` (similar to `Contravariant`'s
@@ -83,9 +80,7 @@ export function getMonoid<R, A>(M: Monoid<A>): Monoid<Reader<R, A>> {
 /**
  * @since 2.0.0
  */
-export const of: <R, A>(a: A) => Reader<R, A> =
-  /*#__PURE__*/
-  ReaderT.of(I.monadIdentity)
+export const of: <R, A>(a: A) => Reader<R, A> = (a) => () => a
 
 // -------------------------------------------------------------------------------------
 // instances
@@ -94,9 +89,7 @@ export const of: <R, A>(a: A) => Reader<R, A> =
 /**
  * @since 2.0.0
  */
-export const map: <A, B>(f: (a: A) => B) => <R>(fa: Reader<R, A>) => Reader<R, B> =
-  /*#__PURE__*/
-  ReaderT.map(I.monadIdentity)
+export const map: <A, B>(f: (a: A) => B) => <R>(fa: Reader<R, A>) => Reader<R, B> = (f) => (fa) => F.flow(fa, I.map(f))
 
 /**
  * @since 3.0.0
@@ -109,9 +102,8 @@ export const functorReader: Functor2<URI> = {
 /**
  * @since 2.0.0
  */
-export const ap: <R, A>(fa: Reader<R, A>) => <B>(fab: Reader<R, (a: A) => B>) => Reader<R, B> =
-  /*#__PURE__*/
-  ReaderT.ap(I.monadIdentity)
+export const ap: <R, A>(fa: Reader<R, A>) => <B>(fab: Reader<R, (a: A) => B>) => Reader<R, B> = (fa) => (fab) => (r) =>
+  F.pipe(fab(r), I.ap(fa(r)))
 
 /**
  * @since 3.0.0
@@ -155,9 +147,11 @@ export const applicativeReader: Applicative2<URI> = {
 /**
  * @since 2.0.0
  */
-export const chain: <A, R, B>(f: (a: A) => Reader<R, B>) => (ma: Reader<R, A>) => Reader<R, B> =
-  /*#__PURE__*/
-  ReaderT.chain(I.monadIdentity)
+export const chain: <A, R, B>(f: (a: A) => Reader<R, B>) => (ma: Reader<R, A>) => Reader<R, B> = (f) => (fa) => (r) =>
+  F.pipe(
+    fa(r),
+    I.chain((a) => f(a)(r))
+  )
 
 /**
  * @since 3.0.0