+++ CHANGELOG.md
@@ -24,6 +24,7 @@ high state of flux, you're at risk of it changing without notice.
   - remove `Choice` module
   - remove `EitherT` module
   - remove `Strong` module
+  - remove `ReaderT` module
   - remove the mega instances in favour of (**b**) splitted instances
   - `Applicative`
     - remove `getApplicativeComposition` in favour of